server/requirements.txt
flask
flask-cors
chromadb
sentence-transformers
supabase
python-dotenv
requests
httpx

